refactor(win): post-package dispatcher for wizard + repackage hooks
